+++ src/dashboard/hooks/use-soap-modal.hook.ts
@@ -14,7 +14,6 @@ export const useSoapModal = (
   const soapModalizeRef = React.useRef<Modalize>(null)
 
   const handleLongPress = () => {
-    console.log('HERE!@')
     soapModalizeRef?.current?.open?.()
   }